+++ src/services/project-folder/index.ts
@@ -1,0 +1,10 @@
+export {
+  BrowserProjectFolderService,
+  DesktopProjectFolderService,
+  PROJECT_FOLDER_FILE_NAME,
+  parseProjectFolderJson,
+  serializeProjectFolderJson,
+  type LocalProjectFolder,
+  type ProjectFolderService,
+} from './ProjectFolderService';
+export { getProjectFolderService } from './getProjectFolderService';